Update protocol parameters to reflect changes

+++ pkg/protocol/engine/consensus/blockgadget/testframework_test.go
@@ -51,7 +51,7 @@ func NewTestFramework(test *testing.T) *TestFramework {
 				VBFactorKey:  10,
 			},
 			TokenSupply:           1_000_0000,
-			GenesisUnixTimestamp:  uint32(time.Now().Truncate(10 * time.Second).Unix()),
+			GenesisUnixTimestamp:  time.Now().Truncate(10 * time.Second).Unix(),
 			SlotDurationInSeconds: 10,
 		},
